[BUG_FIXED] Fix a crash issue (Unicode binary) while close Notepad++ with an RC file opened under Chinese Xp.
git-svn-id: svn://svn.tuxfamily.org/svnroot/notepadplus/repository/trunk@512 f5eea248-9336-0410-98b8-ebc06183d4e3

+++ PowerEditor/src/ScitillaComponent/FindReplaceDlg.cpp
@@ -543,7 +543,7 @@ void Finder::DeleteResult()
 	_MarkingsStruct._length = _pMainMarkings->size();
 
 	assert(_pMainFoundInfos->size() == _pMainMarkings->size());
-	assert(_scintView.execute(SCI_GETLINECOUNT) == _pMainFoundInfos->size() + 1);
+	assert(size_t(_scintView.execute(SCI_GETLINECOUNT)) == _pMainFoundInfos->size() + 1);
 }
 
 void Finder::gotoNextFoundResult(int direction)
